requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
pymupdf==1.27.2.2
pdfplumber==0.11.0
langchain==0.1.20
langchain-text-splitters==0.0.1
sentence-transformers==2.7.0
chromadb==0.5.0
fastapi==0.111.0
uvicorn==0.30.1
groq==0.9.0
python-dotenv==1.0.1
slowapi==0.1.9
